Be pedantic and use u64

+++ src/data_chunks.rs
@@ -13,7 +13,7 @@ pub struct DataChunk {
 struct DataChunkHandle(ffi::duckdb_data_chunk);
 
 impl DataChunk {
-    pub fn new(ty: RawType, column_count: usize) -> Arc<Self> {
+    pub fn new(ty: RawType, column_count: u64) -> Arc<Self> {
         unimplemented!()
     }
     pub fn reset(&self) {
@@ -21,21 +21,20 @@ impl DataChunk {
             ffi::duckdb_data_chunk_reset(self.handle.lock().unwrap().0);
         }
     }
-    pub fn column_count(&self) -> usize {
-        unsafe { ffi::duckdb_data_chunk_get_column_count(self.handle.lock().unwrap().0) as usize }
+    pub fn column_count(&self) -> u64 {
+        unsafe { ffi::duckdb_data_chunk_get_column_count(self.handle.lock().unwrap().0) }
     }
-    pub fn vector(&self, col_idx: usize) -> Arc<Vector> {
+    pub fn vector(&self, col_idx: u64) -> Arc<Vector> {
         unsafe {
-            let v =
-                ffi::duckdb_data_chunk_get_vector(self.handle.lock().unwrap().0, col_idx as u64);
+            let v = ffi::duckdb_data_chunk_get_vector(self.handle.lock().unwrap().0, col_idx);
             Vector::from_raw(v, None)
         }
     }
-    pub fn size(&self) -> usize {
-        unsafe { ffi::duckdb_data_chunk_get_size(self.handle.lock().unwrap().0) as usize }
+    pub fn size(&self) -> u64 {
+        unsafe { ffi::duckdb_data_chunk_get_size(self.handle.lock().unwrap().0) }
     }
-    pub fn set_size(&self, size: usize) {
-        unsafe { ffi::duckdb_data_chunk_set_size(self.handle.lock().unwrap().0, size as u64) }
+    pub fn set_size(&self, size: u64) {
+        unsafe { ffi::duckdb_data_chunk_set_size(self.handle.lock().unwrap().0, size) }
     }
 }
