+++ src/query.rs
@@ -27,17 +27,17 @@ impl QueryResult {
         })
     }
 
-    pub fn column_name(&self, col: usize) -> Option<String> {
+    pub fn column_name(&self, col: u64) -> Option<String> {
         unsafe {
             let p: *const c_char =
-                ffi::duckdb_column_name(self.handle.lock().unwrap().deref_mut(), col as u64);
+                ffi::duckdb_column_name(self.handle.lock().unwrap().deref_mut(), col);
             let nn = NonNull::new(p as *mut c_char)?;
             let cstr = CStr::from_ptr(nn.as_ptr());
             Some(cstr.to_string_lossy().to_string())
         }
     }
 
-    pub fn column_type(&self, col: usize) -> Option<LogicalType> {
+    pub fn column_type(&self, col: u64) -> Option<LogicalType> {
         unsafe {
             LogicalType::from_raw(ffi::duckdb_column_logical_type(
                 self.handle.lock().unwrap().deref_mut(),
@@ -46,14 +46,14 @@ impl QueryResult {
         }
     }
 
-    pub fn column_count(&self) -> usize {
-        unsafe { ffi::duckdb_column_count(self.handle.lock().unwrap().deref_mut()) as usize }
+    pub fn column_count(&self) -> u64 {
+        unsafe { ffi::duckdb_column_count(self.handle.lock().unwrap().deref_mut()) }
     }
-    pub fn row_count(&self) -> usize {
-        unsafe { ffi::duckdb_row_count(self.handle.lock().unwrap().deref_mut()) as usize }
+    pub fn row_count(&self) -> u64 {
+        unsafe { ffi::duckdb_row_count(self.handle.lock().unwrap().deref_mut()) }
     }
-    pub fn rows_changed(&self) -> usize {
-        unsafe { ffi::duckdb_rows_changed(self.handle.lock().unwrap().deref_mut()) as usize }
+    pub fn rows_changed(&self) -> u64 {
+        unsafe { ffi::duckdb_rows_changed(self.handle.lock().unwrap().deref_mut()) }
     }
 }
